Finishing Medical
After receiving some feedback on the changes I've finished up my medical move and touched up cargo somewhat to make it function as intended.

Also gutted medbay from its old location to begin the process of moving Security into its place.

In the words of Nietzsche:
"God is dead. God remains dead. And we have killed him. How shall we comfort ourselves, the murderers of all murderers? What was holiest and mightiest of all that the world has yet owned has bled to death under our knives: who will wipe this blood off us? What water is there for us to clean ourselves? What festivals of atonement, what sacred games shall we have to invent? Is not the greatness of this deed too great for us? Must we ourselves not become gods simply to appear worthy of it?"