Munitions Expansion
Expanded the munitions department up a few tiles to fit in a MAC gun. Players complained that only having the Railgun was lame and made the ship underpowered which is a huge problem with space and player constraints. Hopefully this gives the sausage a little bit of a fighting chance. Next I'll be layering hull plating as required(even though the old map didn't have it?)