Minor Fixes
I found while testing that the brig flashes were not connected properly and the maintenance rooms behind the arcade were a bit crap. I've corrected the issue with the flashes and hopefully also the issue with the maintenance rooms. I highly suspect when this makes it to PR status the arcade will need to be removed.

Next up I'll be trying to get the atmospherics department to look less like absolute chaos and more like something a normal human being can comprehend.