Redundant APC removal
Removes various redundant APCs from a prior error, lowers overall ship energy drain (keep a watch on this, may reduce starting plasma sheets), adds a Pandemic machine, a defib, and an internal lock for the bridge. Also rezones some areas that were supposed to be designated as maintenance to now be as such.