more fun stuff!
Adds a helper proc for checking AI traits.
Looks very big but thats only because it accepts both single elements aswell as lists on the checked side and checking side.

Anyway, in addition to that, Destroyers are now wary of their surroundings, and will investigate when their current target exits their sensor range.
There may be ways to use this to your advantage...