Ship weapon ammo counters, autogauss recipe filepath fix
On examine shows the current ammo count for 50cal turrets, consoles and ammo boxes.
On examine shows the 'powder level' for naval shell loaders, and also adds 'strength' (total shell speed with the extra bag power) and 'shell' (name of the loaded shell type) text data to the UI naval console. Still needs a TGUI change to show it.