fix ghost items on trolleys
Same item could be loaded onto multiple munitions trolleys at once. When it got loaded onto the second one it got removed from the first's contents but not from its vis_contents and the variables for the stacking overlay weren't adjusted, making the item still apear to be on the trolley and taking up space. This fix makes sure that unload_munition gets called by calling it from the Exited proc.