delete Astrometrics Computer's radio on destruction
On construction, the Astrometrics Computer creates a radio in its contents.
All contents get droped on destruction by default. Because the radio
wasn't added in construction this creates it from thin air from the
players perspective. This fix is in line how the same thing is handled for the
SM crystal.
Fixes  #1925