add MAA stamp (#2054)
* take stamp and paper plane icons from nsv folder and add maa stamp

bureaucracy.dmi has been coppied to the nsv folder to prevent merge conflics
in the original bureaucracy.dmi file. The downside of this is that some
upstream additions to that file will need do be copied to the nsv variant
of the file.

* give MAA stamp a hopefully existing item_color

Who knows if it even does something ¯\_(ツ)_/¯

* add //NSV13 comments

Fuck, I keep forgetting those.

* change only MAA stamps icon to use nsv's dmi file

Paper code still uses NSV13's bureaucracy.dmi for all stamps.