remove unneeded CanAtmosPass proc
This proc only gets called if the var with the same name is set to
ATMOS_PASS_PROC. The only parrent that sets that variable is is atom, which
sets it to ATMOS_PASS_YES, meaning this proc never gets called.
There is also the question of whether auxmos even calls that proc.